% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketlist.R
\name{bucketlist}
\alias{bucketlist}
\title{Retrieve S3 Bucket List}
\usage{
bucketlist(region = "us-east-1", access_key = NULL, secret_key = NULL,
  ...)
}
\arguments{
\item{region}{Region of S3 Bucket}

\item{access_key}{Amazon S3 Access Key}

\item{secret_key}{Amazon S3 Secret Key}

\item{...}{arguments to pass to \code{\link{get_hcp_file}}}
}
\value{
List of Buckets
}
\description{
Lists the Buckets available with the Key/Secrete
}
\examples{
\dontrun{
bucketlist()
}
}

