% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hcp_file.R
\name{get_hcp_file}
\alias{get_hcp_file}
\alias{head_hcp_file}
\alias{get_fcp_file}
\alias{head_fcp_file}
\alias{get_openneuro_file}
\alias{head_openneuro_file}
\title{Get HCP file}
\usage{
get_hcp_file(path_to_file = "/", ..., verbose = TRUE, verb = "GET")

head_hcp_file(...)

get_fcp_file(...)

head_fcp_file(...)

get_openneuro_file(...)

head_openneuro_file(...)
}
\arguments{
\item{path_to_file}{Path to file on HCP S3 Bucket}

\item{...}{arguments to pass to \code{\link{make_aws_call}}}

\item{verbose}{Should the URL be printed?}

\item{verb}{passed to \code{\link[httr]{VERB}}}
}
\value{
Result of \code{GET}
}
\description{
Wraps a \code{\link{make_aws_call}} to a \code{\link[httr]{GET}}
statement to get the file
}
