% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_scanning_info.R
\docType{data}
\name{hcp_900_scanning_info}
\alias{hcp_900_scanning_info}
\title{Scanning Information for HCP 900 Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34406 rows and 18 columns.
}
\usage{
hcp_900_scanning_info
}
\description{
A \code{data.frame} with all the available scanning information
for the HCP 900 data.
}
\keyword{datasets}
