% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_ids.R
\name{hcp_ids}
\alias{hcp_ids}
\title{Get list of HCP Ids}
\usage{
hcp_ids(group = c("HCP", "HCP_900", "HCP_1200", "HCP_Retest"))
}
\arguments{
\item{group}{Group of IDs to return}
}
\value{
Character vector
}
\description{
Returns list of IDs of folders in the HCP database
}
\examples{
if (have_aws_key()) {
head(hcp_ids("HCP"))
head(hcp_ids("HCP_900"))
}
}
