% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aws_api_key.R
\name{set_aws_api_key}
\alias{set_aws_api_key}
\alias{have_aws_key}
\title{Set Amazon AWS Key}
\usage{
set_aws_api_key(access_key = NULL, secret_key = NULL,
  default_region = "us-east-1", error = TRUE)

have_aws_key()
}
\arguments{
\item{access_key}{Amazon access key.  If \code{NULL} then looks at the
\code{AWS_ACCESS_KEY_ID} system variable.}

\item{secret_key}{Amazon secret key.  If \code{NULL} then looks at the
\code{AWS_SECRET_ACCESS_KEY} system variable.}

\item{default_region}{Amazon default region.  If \code{NULL} then looks at the
\code{AWS_DEFAULT_REGION} system variable.}

\item{error}{Should this function error if things are not specified?}
}
\value{
List of \code{access_key}, \code{secret_key},
and \code{default_region}
}
\description{
Sets and returns the AWS keys.  This will error if not all are
specified.
}
