% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hcp_file.R
\name{download_hcp_file}
\alias{download_hcp_file}
\title{Download HCP file}
\usage{
download_hcp_file(path_to_file, destfile = NULL, ...)
}
\arguments{
\item{path_to_file}{Path to file on HCP S3 Bucket}

\item{destfile}{Destination filename}

\item{...}{arguments to pass to \code{\link{aws_query_string_auth_url}}}
}
\value{
Output filename that was downloaded
}
\description{
Downloads a file from HCP S3 Bucket
}

