% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hcp_file.R
\name{get_hcp_file}
\alias{get_hcp_file}
\title{Get HCP file}
\usage{
get_hcp_file(..., verbose = TRUE)
}
\arguments{
\item{...}{arguments to pass to \code{\link{hcp_aws_url}}}

\item{verbose}{Should the URL be printed?}
}
\value{
Result of \code{GET}
}
\description{
Wraps a \code{\link{hcp_aws_url}} to a \code{GET}
statement to get the file
}

