% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcp_list_files.R
\name{hcp_list_files}
\alias{hcp_list_dirs}
\alias{hcp_list_files}
\title{List files from HCP bucket}
\usage{
hcp_list_files(prefix = "", delimiter = "", query = NULL, ...)

hcp_list_dirs(prefix = "", delimiter = "/", ...)
}
\arguments{
\item{prefix}{directory folder to list files.  If \code{""}, then it will be
the root path}

\item{delimiter}{Delimiter to list files.  For example \code{".mat"},
for ".mat"}

\item{query}{Additional query arguments}

\item{...}{additional arguments passed to \code{\link{get_hcp_file}}}
}
\value{
List with the result of the GET command, the parsed result, and
the content from the result.
}
\description{
Lists a set of files from an HCP bucket for a specific sub-folder
}
\examples{
if (have_aws_key()){
x = hcp_list_files(prefix = "HCP/100307/unprocessed/3T/Diffusion",
   delimiter="bval")
}

}

